requirements.txt
Flask==2.1.2
requests==2.26.0
gunicorn==20.0.4
PyYAML==6.0
